+++ CHANGELOG.md
@@ -8,6 +8,7 @@ New features:
 - Conversations now move to `.humbug/conversation` instead of `conversations` as they're not directly part of the mindspace
   file content and are auto-managed by Humbug.  The `.humbug` directory now becomes visible to support viewing and drag-and-drop
   operations.
+- Removed the edit button from file previews as editing can now be done via the context menu.
 
 Bug fixes: